move LFPElectrodeGroup to lfp_electrode.py
