+++ src/spyglass/lfp/__init__.py
@@ -1,1 +1,2 @@
 from spyglass.lfp.lfp_merge import LFPOutput
+from spyglass.lfp.lfp_electrode import LFPElectrodeGroup
